Api router

+++ backend/src/v1Api/controllers/shiftsController.ts
@@ -1,41 +1,19 @@
-import express from 'express';
-import prisma from '../../db/db'
-import {io} from '../../src/app'
-import { DateTime } from 'luxon';
-const shift = express.Router();
+import { Request,Response } from "express"
+import prisma from "../../../db/db"
+import {io} from '../../app'
+import { DateTime } from "luxon";
 
 
-shift.post('/:id/shift',async(req,res)=>{
-    const employee_id = parseInt(req.params.id)
-    const {date} = req.body;
-    try{
-        const row = await prisma.employee_Shifts.create({
-           data:{
-                employee_id:employee_id,
-                date: new Date(date),
-                start_time:null,
-                end_time:null
-           }
-        })
 
-        io.emit("shiftAdded",{shift:row})
-        
-        console.log('backend socket emit')
-        res.json({success:true,row})
-    }
-    catch(error:unknown){
-         if(error instanceof Error){
-            console.error('Error',error.message)
-            return res.status(500).json({success:false,error:error.message})
-        }
-      
 
-    }
-})
 
-shift.delete('/:id/delete',async(req,res)=>{
-    const shift_id = parseInt(req.params.id);
+export async function deleteShift(req:Request,res:Response){
+    const shift_id = Number(req.params.id);
    
+    if(Number.isNaN(shift_id)){
+        return res.status(400).json({error:"Invalid shift id"})
+    }
+
     try{
         const deletedShift = await prisma.employee_Shifts.delete({
             where:{id:shift_id},
@@ -50,11 +28,24 @@ shift.delete('/:id/delete',async(req,res)=>{
             return res.status(500).json({success:false,error:error.message})
         }
     }
-})
+}
+export async function moveShift(req:Request, res:Response){
+    const shift_id = Number(req.params.id);
+    const employee_id = Number(req.body.employee_id)
+    const date = new Date(req.body.date);
 
-shift.patch('/moveShift/:id',async(req,res)=>{
-    const shift_id = parseInt(req.params.id);
-    const {employee_id,date} = req.body
+
+    if(Number.isNaN(shift_id)){
+        return res.status(400).json({error:"Invalid shift id"})
+    }
+
+    if(Number.isNaN(employee_id)){
+        return res.status(400).json({error:"Invalid employee id"})
+    }
+    if(!(date instanceof Date) || isNaN(date.getTime())){
+        return res.status(400).json({error:"Invalid date id"})
+    }
+
     try{
        
         const updatedShift = await prisma.employee_Shifts.update({
@@ -63,31 +54,48 @@ shift.patch('/moveShift/:id',async(req,res)=>{
             },
             data:{
                 employee_id:employee_id,
-                date: new Date(date)
+                date: date
             }
         })
-       
 
         io.emit('shiftMoved',{ updatedShift })
         res.status(200).json({ updatedShift });
 
-    }catch(error){
+
+    }catch(error:any){
         console.error("Error updating shift:", error);
-        res.status(500).json({ error: "Failed to update shift" });
+
+        if (error.code === "P2025"){
+            return res.status(404).json({ error: "Shift not found" }); 
+        } 
+        return res.status(500).json({ error: "Failed to update shift" });
     }
 
-})
 
-shift.post('/copyOverLastWeek',async(req,res)=>{
-  
+}
+
+export async function copyOverLastWeek(req:Request,res:Response){
     const {lastWeekArray} = req.body
+
+    if(!Array.isArray(lastWeekArray)){
+        return res.status(400).json({error:"Invalid lastWeekArray"})
+    }
+
+    const validatedArray:Date[] = lastWeekArray.filter((date)=>!DateTime.fromISO(date).isValid);
+
+    if(validatedArray.length > 0){
+        return res.status(400).json({error:"lastWeekArray contains invalid dates"})
+    }
+
+
     const thisWeekArray= lastWeekArray.map((date:string)=>{
         return(
             DateTime.fromISO(date).plus({week:1}).toJSDate()
         )
     })
-
-    await prisma.$transaction(async(tx)=>{
+    console.log(thisWeekArray)
+    try{
+         await prisma.$transaction(async(tx)=>{
 
         // grabs all shifts from last week using last weeks dates sun-sat
         const prevShifts = await tx.employee_Shifts.findMany({
@@ -143,18 +151,10 @@ shift.post('/copyOverLastWeek',async(req,res)=>{
         })
 
         io.emit('copyOverLastWeekshift',data)
-        res.json({success:true,shifts:prevShifts})
+        return res.json({success:true,shifts:newShifts})
     })
-    
-
-    
+    }catch(error:any){
+        return res.status(500).json({ error: "Failed to copy shifts for this week." });
+    }
    
-    
-
-    
-    
-    
-    
-})
-
-export default shift
+}
